refactor: rename data.ts to json.ts
